requirements.txt
pywebview==5.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.32.3
pyinstaller==6.12.0